Fixed null issue
Still crashing sometimes when adding friends and socket is only working
one page at a time. I wanted to have a profile page open, sending its
location back to the server and also have the map page open to show how
it updates. But only one will work at a time, you have to close the page
that is working to get the other to update.